0.1 - Application running, two-page setup
Got the application running properly, with access to the main page and the 'new recipe' page.

Issues:
Creating a new recipe still requires proper setup for ingredient lists. Clicking the 'add ingredient' button does not add the ingredient to the display list, investigate.
Additionally, saving the recipe does not return the correct values to the main page.